fix(trade_manage): 修改Stock_Number结构体中的number类型为double，windows下容易出现精度错误

+++ hikyuu_cpp/hikyuu/trade_manage/TradeManager.cpp
@@ -1262,7 +1262,7 @@ FundsRecord TradeManager::getFunds(const Datetime& indatetime, KQuery::KType kty
         Stock_Number(const Stock& stock, size_t number) : stock(stock), number(number) {}
 
         Stock stock;
-        size_t number;
+        double number;
     };
 
     price_t checkin_cash = 0.0;
